backend/requirements.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.1.0
Flask-Cors==5.0.0
huggingface-hub==0.26.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-rapidjson==1.20
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
sentence-transformers==3.2.1
tokenizers==0.20.1
torch==2.5.0
transformers==4.46.0
urllib3==2.2.3 
Werkzeug==3.1.3
typing_extensions==4.12.2
